+++ R/conversion.R
@@ -6,7 +6,7 @@
 #' @param name_idx column index with sample names. Defaults to NULL. If NULL
 #'                 or 0, rownames(x) or a sequence of integers will be assigned
 #'                 as names.
-#' @param meta_idxs column indices with metadata (not name and not reflectance).
+#' @param meta_idxs column indices with metadata (not name and not value).
 #'                  Defaults to NULL
 #' @return spectra object
 #'
@@ -25,7 +25,7 @@ as.spectra = function(x, name_idx = NULL, meta_idxs = NULL){
 #'
 #' @param x matrix
 #' @param name_idx column index with sample names. Defaults to NULL
-#' @param meta_idxs column indices with metadata (not name and not reflectance).
+#' @param meta_idxs column indices with metadata (not name and not value).
 #'                  Defaults to NULL
 #' @return spectra object
 #'
@@ -55,7 +55,7 @@ as.spectra.matrix = function(x, name_idx = NULL, meta_idxs = NULL){
 #'
 #' @param x data.frame
 #' @param name_idx column index with sample names. Defaults to NULL.
-#' @param meta_idxs column indices with metadata (not name and not reflectance).
+#' @param meta_idxs column indices with metadata (not name and not value).
 #'                  Defaults to NULL
 #' @return spectra object
 #'
@@ -73,7 +73,7 @@ as.spectra.data.frame = function(x, name_idx = NULL, meta_idxs = NULL){
 #' @param fix_names Use make.names to normalize names?
 #'                  Pick one: "none" "row" "col" "both".
 #' @param ... does nothing. Here for compatibility with S3 generics
-#' @return matrix of spectral reflectance. columns are wavelengths and rows are
+#' @return matrix of spectral value. columns are wavelengths and rows are
 #'         samples
 #'
 #' @author Jose Eduardo Meireles
@@ -84,7 +84,7 @@ as.spectra.data.frame = function(x, name_idx = NULL, meta_idxs = NULL){
 #' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' mat  = as.matrix(spec)
 as.matrix.spectra = function(x, fix_names = "none", ...) {
-    r = reflectance(x)
+    r = value(x)
     s = names(x)
     w = wavelengths(x)
     o = c("none", "row", "col", "both")
@@ -108,7 +108,7 @@ as.matrix.spectra = function(x, fix_names = "none", ...) {
 #' Convert spectra to data.frame
 #'
 #' Returns a data.frame that includes sample names, metadata (if present) and
-#' reflectance data. One advantage over as.matrix, is that the metadata are
+#' value data. One advantage over as.matrix, is that the metadata are
 #' returned.
 #'
 #' @param x spectra object
@@ -118,7 +118,7 @@ as.matrix.spectra = function(x, fix_names = "none", ...) {
 #'                  Pick one: "none" "row" "col" "both".
 #' @param metadata boolean. Include spectral metadata? Defaults to TRUE
 #' @param ... extra parameters passed to the generic as.spectra
-#' @return data.frame with: sample_name, metadata (if any) and reflectance.
+#' @return data.frame with: sample_name, metadata (if any) and value.
 #'
 #' @author Jose Eduardo Meireles
 #' @export