+++ R/match_sensors.R
@@ -109,7 +109,7 @@ i_trim_sensor_overlap = function(x, splice_at){
 
 #' Match spectra at sensor transitions
 #'
-#' \code{match_sensors} scales reflectance values of sensors 1 (vis) and 3 (swir2)
+#' \code{match_sensors} scales value values of sensors 1 (vis) and 3 (swir2)
 #'
 #' Splice_at has no default because sensor transition points vary between vendors
 #' and individual instruments. It is an important parameter though, so you should
@@ -120,7 +120,7 @@ i_trim_sensor_overlap = function(x, splice_at){
 #'
 #' If the factors used to match spectra are unreasonable, \code{match_sensors}
 #' will throw. Unreasonable factors (f) are defined as 0.5 > f > 3 or NaN,
-#' which  happens when the reflectance value for the right sensor is 0.
+#' which  happens when the value value for the right sensor is 0.
 #'
 #' @param x spectra object
 #' @param splice_at wavelengths that serve as splice points, i.e the beginnings
@@ -135,12 +135,6 @@ i_trim_sensor_overlap = function(x, splice_at){
 #'
 #' @author Jose Eduardo Meireles and Anna Schweiger
 #' @export
-#'
-#' @examples
-#' library(spectrolab)
-#' data(spec_with_jump)
-#' spec = spec_with_jump
-#' spec = match_sensors(spec, splice_at = c(971, 1910), fixed_sensor = 2)
 match_sensors = function(x,
                          splice_at,
                          fixed_sensor    = 2,
@@ -225,8 +219,8 @@ match_sensors.spectra = function(x,
             scaled = wl_picks[[z]]$left
         }
 
-        rowMeans(reflectance(x[ , scaled, simplify = FALSE])) /
-        rowMeans(reflectance(x[ ,  fixed, simplify = FALSE]))
+        rowMeans(value(x[ , scaled, simplify = FALSE])) /
+        rowMeans(value(x[ ,  fixed, simplify = FALSE]))
     })
 
     ## Verify if factors for splicing are reasonable
@@ -255,7 +249,7 @@ match_sensors.spectra = function(x,
 
     ## Transform data
     for(i in seq_along(factor_mat)){
-        x[ , s[[i]]] = reflectance(x[ , s[[i]] ] ) * t( factor_mat[[i]] )
+        x[ , s[[i]]] = value(x[ , s[[i]] ] ) * t( factor_mat[[i]] )
     }
 
     x