+++ NEWS.md
@@ -9,7 +9,7 @@
 
 ## major
 * Fixed major bug. Spectrum replacement (e.g. spec_a[i, ] = spec_b[1, ]) only
-  replaced the reflectance and not the other information (names, metadata, etc.)
+  replaced the value and not the other information (names, metadata, etc.)
 ## minor
 * `quantile` now allows the user to choose names for the spectra
 
@@ -27,7 +27,7 @@
 # spectrolab 0.0.5 (2018-05-31)
 
 ## major
-* Fixed bug in "reflectance" setter, where the object was not being returned
+* Fixed bug in "value" setter, where the object was not being returned
 * The parser inside read_spectra now finds the spectral data using tags instead
   of assuming that the data starts at a specific line